+++ roleplay/unicornia/images/__init__.py
@@ -1,2 +1,0 @@
-from .images import *  # noqa: F403
-from .text import *  # noqa: F403